feat(aot): align moe precompile with runtime wrappers
